requirements.txt
Flask==2.3.2
google-genai==1.0.0
python-dotenv==1.0.0
Flask-Session==0.5.0